fix: logging
Logging is now setup after loading and validating settings from
config.ini
A small change to current logging utility has been made with respect to
request_id, if the value aint in the context. it probably means that the
logger was called from outside the wsgi __call__. So that is specified
in the log, if thats the case.