Bug Fix - block put location specification
Considered using REGEX, however that would mean doing a REGEX search/replace which would not be very easy to track, etc. However, the simpler solution is to just breakdown the URL, remove/update the parts we need to, and then rebuild it.

There is also code there to be enabled once PR #173 (differing block ids between metadata and storage) is merged; it might need a little tweaking though but should be near 100% correct and just need to be uncommented