Cleanup - imorts
In trying to test Falcon master to fix the issue with no body in the 405 reponse I discovered we had an import that Falcon removed (falcon.api_helpers) in latest master.
Only three files imported it, and none needed it. cleaned up the imports so that these three are using their minimal imports. We should now be compatible with Falcon master branch too.