Enhancement: More explicit and accurate error message when underlying storage block does not exist
This converts Deuce from the 502/Bad Gateway error over to the 410/Gone error

Previously 502/Bad Gateway was being returned under the argument that Deuce is a "gateway" to storage in the normal
deployment use-case (e.g OpenStack Swift/Rackspace Cloud Files). However, it's not really a service error as the 5xx
error codes are reserved for. After discussion the team agrees on 410/Gone instead; client are suppose to be able
remedy this by re-uploading the block which will then restore everything dependent on the missing data.